Time: 46 ms (87.62%), Space: 39.6 MB (58.48%) - LeetHub

+++ 1191-k-concatenation-maximum-sum/1191-k-concatenation-maximum-sum.cpp
@@ -29,7 +29,7 @@ public:
             return maxsub;
         }
         else{
-            long long int ans = (maxsub%MOD*1LL + (1LL*sum*(k-2))%MOD)%MOD;
+            int ans = (maxsub%MOD*1LL + (1LL*sum*(k-2))%MOD)%MOD;
             return ans;
         }
         return 0;
